requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.31.0
pypinyin==0.49.0
